% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webterminal-package.R
\docType{package}
\name{webterminal-package}
\alias{webterminal}
\alias{webterminal-package}
\title{webterminal: Fast Browser-Based Terminal for 'RStudio Server'}
\description{
Opens a browser-based terminal inside the 'RStudio Server' viewer pane by forwarding to a locally running 'ttyd' or 'shellinabox' daemon and an optional 'tmux' session. The package does not install or modify any system software; it only locates user-installed helper binaries, manages daemon lifecycle, and proxies through 'RStudio Server' built-in localhost URL proxy via 'rstudioapi'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sondreskarsten/webterminal}
  \item Report bugs at \url{https://github.com/sondreskarsten/webterminal/issues}
}

}
\author{
\strong{Maintainer}: Sondre N<c3><a6>r<c3><b8> Skarsten \email{Sondre.Naro.Skarsten@dnb.no} [copyright holder]

Authors:
\itemize{
  \item Sondre N<c3><a6>r<c3><b8> Skarsten \email{Sondre.Naro.Skarsten@dnb.no} [copyright holder]
}

}
\keyword{internal}
